[level 1] Title: 부족한 금액 계산하기, Time: 0.01 ms, Memory: 4.11 MB -BaekjoonHub

+++ 프로그래머스/1/82612. 부족한 금액 계산하기/부족한 금액 계산하기.cpp
@@ -3,7 +3,7 @@ using namespace std;
 long long solution(int price, int money, int count)
 {
     long long answer = 0;
-    answer = (long long)money;
+    answer = money;
     
     for (int i = 1 ; i <= count ; i++)
     {
